/pkg/declarative: Move to /internal
