+++ README.md
@@ -94,7 +94,7 @@ Here is a (somewhat complete) list of the different modules in the system togeth
 | v1alpha1,v1beta1 | [Manifest](api/v1beta1/manifest_types.go)                 | Beta-Grade - no breaking changes without API incrementation. Use for automation and watch upstream as close as possible for deprecations or new versions. Alpha API is deprecated and converted via webhook. |
 | v1alpha1,v1beta1 | [Watcher](api/v1beta1/watcher_types.go)                   | Beta-Grade - no breaking changes without API incrementation. Use for automation and watch upstream as close as possible for deprecations or new versions. Alpha API is deprecated and converted via webhook. |
 |                  | [Kyma Controller](controllers/kyma_controller.go)         | In active development (continuous) - Expect Bugs and fast-paced development of new features                                                                                                                  |
-|                  | [Manifest Controller](controllers/manifest_controller.go) | Directs to the [Declarative Library](pkg/declarative/v2), a reconciliation library we use to install all modules                                                                                             |
+|                  | [Manifest Controller](controllers/manifest_controller.go) | Directs to the [Declarative Library](internal/declarative/v2), a reconciliation library we use to install all modules                                                                                             |
 |                  | [Watcher Controller](controllers/watcher_controller.go)   | Maintains VirtualService entries for events coming from runtime clusters, mostly stable                                                                                                                      |
 
 ## Deployment / Delivery models
