+++ api/v1beta1/manifest_types.go
@@ -17,7 +17,7 @@ limitations under the License.
 package v1beta1
 
 import (
-	declarative "github.com/kyma-project/lifecycle-manager/pkg/declarative/v2"
+	declarative "github.com/kyma-project/lifecycle-manager/internal/declarative/v2"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
