+++ internal/declarative/v2/crds.go
