+++ internal/declarative/v2/inmemory_rendered.go
