+++ internal/declarative/v2/object.go
