+++ internal/declarative/v2/options.go
