+++ internal/declarative/v2/raw_render.go
