+++ internal/declarative/v2/ready_check.go
