+++ internal/declarative/v2/renderer_with_cache.go
