+++ internal/declarative/v2/renderer_with_cache_test.go
@@ -3,11 +3,11 @@ package v2_test
 import (
 	"context"
 	"errors"
+	. "github.com/kyma-project/lifecycle-manager/internal/declarative/v2"
+	mockV2 "github.com/kyma-project/lifecycle-manager/internal/declarative/v2/mock"
 	"testing"
 
 	"github.com/golang/mock/gomock"
-	. "github.com/kyma-project/lifecycle-manager/pkg/declarative/v2"
-	mockV2 "github.com/kyma-project/lifecycle-manager/pkg/declarative/v2/mock"
 	"github.com/stretchr/testify/assert"
 	"k8s.io/client-go/tools/record"
 )
