+++ internal/declarative/v2/rest_getter_proxy.go
