+++ internal/declarative/v2/spec.go
