+++ internal/declarative/v2/ssa.go
