+++ internal/declarative/v2/test/v1/suite_test.go
@@ -2,13 +2,11 @@ package v1_test
 
 import (
 	"context"
+	testv1 "github.com/kyma-project/lifecycle-manager/internal/declarative/v2/test/v1"
 	"os"
 	"path/filepath"
 	"time"
 
-	testv1 "github.com/kyma-project/lifecycle-manager/pkg/declarative/v2/test/v1"
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
 	apiExtensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"k8s.io/client-go/kubernetes/scheme"
 
@@ -17,6 +15,9 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 	"sigs.k8s.io/yaml"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
 )
 
 const (
