+++ internal/declarative/v2/util.go
